+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [652](https://github.com/MiKTeX/miktex/issues/652): miktexsetup.exe: MiKTeX encountered an internal error.
 - [659](https://github.com/MiKTeX/miktex/issues/659): Check for directory that executable doesn't have permissions for causes crash
 - [666](https://github.com/MiKTeX/miktex/issues/666): miktex (lualatex.exe) fails to run from flask subprocess
+- [668](https://github.com/MiKTeX/miktex/issues/668): expand tilde prefixes
 
 ## 20.11 - 2020-11-02